fix(gui): fix build in GitHub Actions
