+++ jadx-gui/src/main/java/jadx/gui/ui/DnDTabData.java
@@ -1,0 +1,9 @@
+package jadx.gui.ui;
+
+class DnDTabData {
+	public final DnDTabbedPane tabbedPane;
+
+	protected DnDTabData(DnDTabbedPane tabbedPane) {
+		this.tabbedPane = tabbedPane;
+	}
+}
